Add stratumweight parameter
Currently, for RedHat and Debian, the module has `stratumweight`
hardcoded to `0`.  Chrony's default has been `0.001` since version 2.0 and
the default configuration file shipped with EL7 doesn't set it to `0`
either. (I suspect a similar situation with modern Debian OSes).

The new parameter defaults to `undef` and leaves out explicitly
configuring `stratumweight` in the configuration file.
This is technically a breaking change as to preserve the
old behaviour, it has to be set to `0`.  I still think this is
preferable to defaulting the parameter to `0` (for OSes other than ArchLinux)
and expecting users to know that `0.001` would be a better choice.