Add parameter dumpdir.
To be able to specify directory for chrony to save measurement in on exit. When dumpdir is
set, both option dumponexit and dumpdir is added to chrony.conf

dumponexit and dumpdir was before b971db06012b1104bbf5bd17382a573e6f354b3a hardcoded in
the template for archlinux.  This commit is restoring the default value for Archlinux, and
leaving the default unset on other os.