Fix template to not add trailing space after servers,peers,pools
The attempt to simplify the template by letting flatten() take care of the different
allowed data-types for servers,peers, and pools, was unfortunately a bit too
optimistic. The result was that join(' ') on a Array with empty element created a extra
space caracter when $chrony::servers was a hash with a key but no value.

This commit uses filter to remove empty elements created by flatten before join.