enforce Array[String] data type for queryhosts
The conversion from ERB to EPP in #79 changed the behavior of
queryhosts slightly. Before, setting queryhosts to the empty string
worked for configuring chrony to allow any query. After, empty string
is considered empty, so the "allow" directive would not be rendered.

Enforcing the data type ensures that people upgrading this module who
have been using empty string get an error rather than a valid but
misconfigured chrony.conf. The replacement for empty string is an
array with a single empty string element: ['']