Allow users to not set local stratum
The template uses a conditional statement but it always evaluates to true
and a result users can't disable the usage of local stratum.

We can fix that by adjusting the data type of the local_stratum
parameter to accept the Boolean false value. So, users can set it to
`false` and the template doesn't render the specific block as the
conditional statement evaluates to false.