Fix CI on CentOS
We are experiencing CI failures on CentOS because the service cannot
start because it cannot create it's PID file.

Running the very same test locally works as expected, so the root cause
of the different behavior on the CI is unfortunately unknown.  Yet,
ensuring the directory for the PID file exist seems to fix the CI issue.

So manage explicitely this directory when running on CentOS in CI.