make config_keys option optional
Previously we always wrote `keyfile` to the chrony configuration file.
This is not always desired. To not cause a breaking change but still
make this configureable, the parameter is now a variant. Besides passing
an absolute path, you can now set it to `''` (empty string). If that's
the case, the keys file won't be managed by Puppet and the option will
be removed from the config file.