init: queryhosts and denyqueryhosts should accept empty strings
Commit ee2d9de0b4a8b596169db1a8ff15995978661e44 changed the data type
for queryhosts from Array[String] to Array[String[1]]. These
parameters configure `allow` and `deny` in chrony.conf. But `allow` and
`deny` are allowed to be by themselves in the config, meaning "allow
everything" or "deny everything": https://chrony.tuxfamily.org/doc/4.3/chrony.conf.html#allow

This commit corrects the data type to Array[String[0]] and adds a test
to prevent future breakage (this was broken before too, and I fixed it
again in #101).