fix refclock generation
This updates the refclock parameter data types to be a hash which
matches the examples and updates the template to properly generate
multiple refclock entries.

This looks to have broken via a combination of #141 and #79. Rather than
restore the full variations that were supported prior to #79 a simplified
hash structure was chosen.

Fixes #189

Signed-off-by: Ben Magistro <koncept1@gmail.com>